Seperate LibTable class from check_lib_table functionality
This makes re-using the LibTable class a bit easier and logical. Could be used
to e.g check custom options in LibTables.

It also remove schematic lib specific code from LibTable making it re-usable
for footprint libraries.
